Remove `@module` comments
We are reexporting items to the top-level, instead of expositing them under sub-modules like `antithesis-sdk/assert`. So those `@module` comments can leads to confusing JSDoc items.
And `tsc` does not seem to play well with them anyway, `@module` comments are stripped out for `assert` & `lifecycle` for some reason, and only the one for `random` is preserved...